Changed variable type

+++ ktrigger.h
@@ -13,6 +13,6 @@ struct config_s {
 void usage(char*);
 void init_config(struct config_s*);
 void parse_cmd(struct config_s*, int, char**);
-int run_trigger(char*, char*, int);
+int run_trigger(char*, char*, short);
 
 #endif	/* KTRIGGER_H */
